Changed hash-b to meet the bonus criteria:
(BONUS) For the brave of heart - figure out a method using the second program to print out the months
(and print the holidays as well) in order they were input.

Also changed the for loop to a while loop so it will loop until the user enters "done"
and removed the check for a duplicate hash key since we will not need it with the
use of $count as the key.